% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adj.ind}
\alias{adj.ind}
\title{Adjacency matrix for individuals}
\usage{
adj.ind(rel)
}
\arguments{
\item{rel}{an affiliation list}
}
\value{
a sparse adjacency matrix of individual * individual
}
\description{
Create an adjacency matrix from an affilation list
}

