% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adj.org}
\alias{adj.org}
\title{Adjacency matrix for affiliations}
\usage{
adj.org(rel)
}
\arguments{
\item{rel}{an affiliation list}
}
\value{
a sparse adjacency matrix of affiliation * affiliation
}
\description{
Create an adjacency matrix from an affilation list
}
\seealso{
\link{adj.ind}, \link{two.mode}
}

