% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beskrivelser}
\alias{beskrivelser}
\title{Combine descriptions}
\usage{
beskrivelser(x, rel.all = rel.all)
}
\arguments{
\item{x}{the name of an individual}

\item{rel.all}{an affiliation edge list}
}
\value{
a character vector
}
\description{
Combine all descriptions into a single character vector
}

