% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{description.vector}
\alias{description.vector}
\title{A vector of descriptions for each individual}
\usage{
description.vector(x, den)
}
\arguments{
\item{x}{a character vector with names of individuals}

\item{den}{}
}
\value{
a vector
}
\description{
A vector of descriptions for each individual
}

