% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{edge.coord}
\alias{edge.coord}
\title{Edge coordinates}
\usage{
edge.coord(graph, layout)
}
\arguments{
\item{graph}{a \link{igraph} network object}

\item{layout}{a two-column numerical matrix with coordinates for each vertex of graph}
}
\value{
an edge list as a matrix with coordinates of the start and end vertices for each edge.
}
\description{
Generates coordinates as a data.frame suitable for ggplot2's \link{geom_segment}. This function is not super efficient and should be upgraded.
}

