% Generated by roxygen2 (4.0.2): do not edit by hand
\name{efternavne}
\alias{efternavne}
\title{Extract last names}
\usage{
efternavne(navne)
}
\arguments{
\item{navne}{a character vector of full nmaes}
}
\value{
a character vector of last names
}
\description{
Extract last names from full names
}

