% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reports.R
\name{ego.network}
\alias{ego.network}
\title{Ego network}
\usage{
ego.network(name, den, n = Inf)
}
\arguments{
\item{name}{}

\item{den}{}

\item{n}{}
}
\value{
a ggplot2 plot
}
\description{
An ego network of individuals
}

