% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reports.R
\name{ego.network.two.mode}
\alias{ego.network.two.mode}
\title{A report of the first neighborhood of an individiual}
\usage{
ego.network.two.mode(name, den, n = Inf)
}
\arguments{
\item{name}{a single or several names}

\item{den}{an affiliation edge list in the \link{den} format}

\item{the}{maximum social distance - or highest edge weight - that a person can be from the individuals in name}
}
\value{
a ggplot ego.plot
}
\description{
A report of the first neighborhood of an individiual
}

