% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{ego.two.mode.affil}
\alias{ego.two.mode.affil}
\title{A report of the first neighborhood of an affiliation}
\usage{
ego.two.mode.affil(name, den = den, text = "affil", member.of = pe13$Name,
  text.background = "white", ...)
}
\arguments{
\item{name}{a single or several affilation names}

\item{den}{an affiliation edge list in the \link{den} format}

\item{text}{if "affil" only affiliations are given labels}

\item{member.of}{is a set of names of individuals or affiliations that are used for determining the fill of the vertices.}

\item{...}{all other arguments are passed on to \link{graph.plot.twomode}}
}
\value{
a ggplot ego.plot
}
\description{
A report of the first neighborhood of an affiliation
}

