% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reports.R
\name{ego.two.mode}
\alias{ego.two.mode}
\title{A report of the first neighborhood of an individiual}
\usage{
ego.two.mode(name, den = den, n = Inf, text = "affil",
  member.of = pe13$Name, ...)
}
\arguments{
\item{name}{a single or several names}

\item{den}{an affiliation edge list in the \link{den} format}

\item{n}{the maximum social distance - or highest edge weight - that a person can be from the individuals in name}

\item{text}{if "affil" only affiliations are given labels}

\item{member.of}{is a set of names of individuals or affiliations that are used for determining the fill of the vertices.}

\item{...}{all other arguments are passed on to \link{graph.plot.twomode}}
}
\value{
a ggplot ego.plot
}
\description{
A report of the first neighborhood of an individiual
}

