% Generated by roxygen2 (4.0.1): do not edit by hand
\name{elite.network}
\alias{elite.network}
\title{Elite network}
\usage{
elite.network(rel.all = rel.all, sigma = 14)
}
\arguments{
\item{rel.all}{an affiliation edge list}

\item{sigma}{the number of members in an affiliation above which all affiliations are weighted down}
}
\value{
a elite network object
}
\description{
Construct a weighted elite network
}

