% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find.core.net}
\alias{find.core.net}
\title{Find the core in an elite network - using a graph as input}
\usage{
find.core.net(net, reach = 2.1)
}
\arguments{
\item{net}{an \link{igraph} network object}

\item{reach}{the maximum distance considered as a relation in the decompostition}
}
\value{
a numerical vector with the coreness score for each vertex
}
\description{
Uses the k-core decomposition, see \link{graph.adjacency}, to identify the elite
}

