% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis.R
\name{find.core}
\alias{find.core}
\title{Find the core in an elite network}
\usage{
find.core(sp, reach = 2.1)
}
\arguments{
\item{sp}{a shortest paths matrix}

\item{reach}{the maximum distance considered as a relation in the decompostition}
}
\value{
a numerical vector with the coreness score for each vertex
}
\description{
Uses the k-core decomposition, see \link{graph.adjacency}, to identify the elite
}
\examples{
data(den)
health.affil  <- has.tags(den, c("Health"))
den.health    <- den[den$AFFILIATION \%in\% health.affil,]
adj.health    <- adj.ind(den.health)
net.health    <- graph.adjacency(adj.health)
net.health    <- graph.adjacency(adj.health, mode = "undirected", weighted = TRUE)
sp.health     <- shortest.paths(net.health)
core.health   <- find.core(sp.health)
table(core.health)
}

