% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{fornavne}
\alias{fornavne}
\title{Extract first names}
\usage{
fornavne(navne)
}
\arguments{
\item{navne}{a character vector of full nmaes}
}
\value{
a character vector of first names
}
\description{
Extract first names from full names
}

