% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{graph.plot.twomode}
\alias{graph.plot.twomode}
\title{Twomode graphs}
\usage{
graph.plot.twomode(graph, layout = "default", vertex.fill = "type",
  vertex.size = "degree", edge.color = "edge.betweenness",
  edge.alpha = "edge.betweenness", ...)
}
\arguments{
\item{graph}{a twomode graph}

\item{layout}{if "default"}
}
\value{
a ggplot2 plot
}
\description{
Twomode graphs
}

