% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{hvad}
\alias{hvad}
\title{Hvad er det?}
\usage{
hvad(affil, den = "den", ignore.case = TRUE, tags = FALSE, ...)
}
\arguments{
\item{affil}{a character string with the name of one or several affiliations}

\item{den}{an affiliation edge list in a similar format to \link{den}, if "den" the den dataset is used.}

\item{ignore.case}{if TRUE grep is not case sensitive}

\item{...}{further arguments are passed on to \link{grep}}
}
\value{
A matrix with names and affiliation
}
\description{
Returns the list of members of the affiliations. Names are matched via grep.
}

