% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis.R
\name{hvem}
\alias{hvem}
\title{Hvem er det?}
\usage{
hvem(name = NULL, relation.matrix = rel.all, memberships = TRUE)
}
\arguments{
\item{name}{the name of the individual}

\item{relation.matrix}{a affiliation edge list}

\item{memberships}{if FALSE returns all direct contacts}
}
\description{
Returns the affiliation memberships of an individual, or all direct contacts
}

