% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inddel.postnummer}
\alias{inddel.postnummer}
\title{Categories from postal codes}
\usage{
inddel.postnummer(x)
}
\arguments{
\item{x}{a numeric vector with 4 digit danish postal codes}
}
\value{
a data.frame with various factors
}
\description{
Categories from postal codes
}

