% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{membership.vector}
\alias{membership.vector}
\title{A vector of memberships for each individual}
\usage{
membership.vector(x, den)
}
\arguments{
\item{x}{a character vector with names of individuals}

\item{den}{}
}
\value{
a vector
}
\description{
A vector of memberships for each individual
}

