% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis.R
\name{network.by.variable}
\alias{network.by.variable}
\title{Network by variable}
\usage{
network.by.variable(graph, variabel)
}
\arguments{
\item{graph}{is a \link{igraph} network}

\item{variabel}{is a factor of the same length and order as the vertices in graph}
}
\value{
a matrix with descriptives
}
\description{
Splits a network by a variable and returns matrix of descriptive values
}

