% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis.R
\name{proximity}
\alias{proximity}
\title{Social proximity}
\usage{
proximity(graph, neighborhood = 2, mode = "total")
}
\arguments{
\item{graph}{is a \link{igraph} network}

\item{mode}{if "total" the proximity is calculated on the size of the combined neighborhood. If "own" or "other" proximity is calculated on the basis of either of the vertices in a relation.}

\item{neihborhood}{a numerical value indicating the order of the neighborhood, see \link{neighborhood}}
}
\value{
a matrix with proximity measures
}
\description{
Calculates the social proximity of all vertices in a graph as described by Alba and Kadushin
}

