% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis.R
\name{secondary.actors}
\alias{secondary.actors}
\title{Secondary actors}
\usage{
secondary.actors(x, rel.all)
}
\arguments{
\item{x}{a named core numerical vector with coreness values, see \link{graph.coreness}}

\item{rel.all}{an affiliation edge list}
}
\value{
something
}
\description{
Identify secondary actors within a group. A secondary actor is an individual with a neighborhood that is perfectly nested within another individuals neighborhood.
Here it is identified by comparing memberships between all agents within a group. If any individual has the exact same memberships as another individual he is considered a secondary actor.
}

