% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_network.R
\name{show.all.tags}
\alias{show.all.tags}
\title{show.all.tags}
\usage{
show.all.tags(den)
}
\arguments{
\item{den}{a affiliation edgelist}
}
\description{
Displays a matrix with all the tags in den.
}

