% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tag.network}
\alias{tag.network}
\title{Tag network}
\usage{
tag.network(rel.tag, plot = TRUE, link.col = "ORG_NAVN")
}
\arguments{
\item{rel.tag}{an affiliation list}

\item{plot}{If TRUE the network is plotted}

\item{link.col}{is the column name in the affiliation list that is used as the edges}
}
\value{
a graph object or a plot
}
\description{
Creates and plots a tag network from an affilation list
}

