% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{tag.relations}
\alias{tag.relations}
\title{Tag relation}
\usage{
tag.relations(rel.all, org.names, by.individual = FALSE)
}
\arguments{
\item{rel.all}{}
}
\description{
Finds relations for each affiliation
}

