% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{tag.vector}
\alias{tag.vector}
\title{Create a vector of tags on the basis of a affiliations}
\usage{
tag.vector(x, den)
}
\arguments{
\item{x}{a character vector of affiliation names}

\item{den}{}
}
\value{
a vector of tags
}
\description{
Create a vector of tags on the basis of a affiliations
}

