% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tile.plot}
\alias{tile.plot}
\title{Tile.plot
Plots a matrix as tile with color according to intensity}
\usage{
tile.plot(adj.mat, max.color.value = median(adj.mat) * 2)
}
\arguments{
\item{adj.mat}{is the input matrix, with named rows and columns and numerical cells}
}
\value{
a ggplot2 tile plot
}
\description{
Tile.plot
Plots a matrix as tile with color according to intensity
}

