% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{variable.from.rel.org}
\alias{variable.from.rel.org}
\title{Create an organisation variable from relations matrix}
\usage{
variable.from.rel.org(rel, net, var)
}
\arguments{
\item{rel}{is a relations matrix - or a affiliation edge list}

\item{net}{is an igraph network object}

\item{var}{is a variable of the same length and order as rel - often it would be "SOURCE" or "TAG"}
}
\value{
a character vector
}
\description{
Create an organisation variable from relations matrix
}

