% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{vertex.coord}
\alias{vertex.coord}
\title{Vertex.coord}
\usage{
vertex.coord(graph, layout = layout.fruchterman.reingold(graph))
}
\arguments{
\item{graph}{a \link{igraph} network object}

\item{layout}{a two-column numerical matrix with coordinates for each vertex of graph}
}
\value{
a data.frame with vertex coordinates
}
\description{
Convienience funcion for streamlining vertex.coordinates
}

