% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{vertex.measures.directed}
\alias{vertex.measures.directed}
\title{Vertex descriptives for directed graphs}
\usage{
vertex.measures.directed(net, n = 2.5)
}
\arguments{
\item{net}{is an directed \link{igraph} network object}

\item{reach}{is the maximum distance between two individuals for the reach statistic}
}
\value{
a matrix with a lot of descriptives
}
\description{
Descriptive statistics for each vertex
}

