% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis.R
\name{who}
\alias{who}
\title{Who is it?}
\usage{
who(net, name = NULL, relation.matrix = rel, vertex = NULL)
}
\arguments{
\item{net}{a \link{igraph} network object}

\item{name}{the name of the individual}

\item{relation.matrix}{a affiliation edge list}

\item{vertex}{the index number of the individual}
}
\description{
Affiliation and descriptives for an individual
}

