% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{adj.ind}
\alias{adj.ind}
\title{Adjacency matrix for individuals}
\usage{
adj.ind(rel)
}
\arguments{
\item{rel}{an affiliation list}
}
\value{
a sparse adjacency matrix of individual * individual
}
\description{
Create an adjacency matrix from an affilation list
}

