% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{efternavne}
\alias{efternavne}
\title{Extract last names}
\usage{
efternavne(navne)
}
\arguments{
\item{navne}{a character vector of full nmaes}
}
\value{
a character vector of last names
}
\description{
Extract last names from full names
}

