% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis.R
\name{elite.network.org}
\alias{elite.network.org}
\title{Elite network for affiliations}
\usage{
elite.network.org(rel.all = rel.all, sigma = 14)
}
\arguments{
\item{rel.all}{an affiliation edge list}

\item{sigma}{the number of members in an affiliation above which all affiliations are weighted down}
}
\value{
a elite network object
}
\description{
Construct a weighted elite network of affiliations
}

