% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find.core}
\alias{find.core}
\title{Find the core in an elite network}
\usage{
find.core(sp, reach = 2.1)
}
\arguments{
\item{sp}{a shortest paths matrix}

\item{reach}{the maximum distance considered as a relation in the decompostition}
}
\value{
a numerical vector with the coreness score for each vertex
}
\description{
Uses the k-core decomposition, see \link{graph.adjacency}, to identify the elite
}

