% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{find.gender}
\alias{find.gender}
\title{Find the gender by name}
\usage{
find.gender(navne)
}
\arguments{
\item{navne}{a character vector of full names}

\item{names.gender}{a matrix with national distributions of first names}
}
\value{
a factor with a gender guess
}
\description{
Guesses the gender for a list of names by comparing it to the national distribution of first names.
}

