% Generated by roxygen2 (4.0.2): do not edit by hand
\name{graph.plot}
\alias{graph.plot}
\title{graph.plot - Plots a weighted graph, with weighted edges as a default}
\usage{
graph.plot(graph, layout = layout.fruchterman.reingold(graph, weight =
  E(graph)$weight^2), vertex.color = "black", vertex.fill = "grey60",
  vertex.shape = 21, vertex.size = 3, vertex.alpha = 1, edges = TRUE,
  edge.color = "darkblue", edge.alpha = E(graph)$weight, edge.size = 1,
  edge.line = "solid", edge.order = FALSE, text = FALSE, text.size = 3,
  text.color = "black", text.alpha = 1, legend = "side",
  text.vjust = 1.5, midpoints = FALSE, midpoint.arrow = arrow(angle = 20,
  length = unit(0.33, "cm"), ends = "last", type = "closed"),
  edge.text = FALSE, edge.text.size = 3, edge.text.alpha = 0.9)
}
\arguments{
\item{graph}{a \link{igraph} network object}

\item{layout}{a two-column numerical matrix with coordinates for each vertex of graph}

\item{vertex.color}{}

\item{vertex.fill}{}

\item{vertex.shape}{}

\item{vertex.size}{}

\item{vertex.alpha}{}

\item{edges}{}

\item{edge.color}{}

\item{edge.alpha}{}

\item{edge.size}{}

\item{edge.line}{}

\item{edge.order}{}

\item{text}{}

\item{text.size}{}

\item{text.color}{}

\item{text.alpha}{}

\item{legend}{}

\item{text.vjust}{}

\item{midpoints}{}

\item{midpoint.arrow}{}

\item{edge.text}{}

\item{edge.text.size}{}

\item{edge.text.alpha}{}
}
\description{
graph.plot - Plots a weighted graph, with weighted edges as a default
}

