% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reports.R
\name{graph.plot.twomode}
\alias{graph.plot.twomode}
\title{Twomode graphs}
\usage{
graph.plot.twomode(graph, layout = "default", vertex.fill = "type",
  vertex.size = "degree", edge.color = "edge.betweenness",
  edge.alpha = "edge.betweenness", ...)
}
\arguments{
\item{graph}{}

\item{layout}{}

\item{vertex.fill}{}

\item{vertex.size}{}

\item{edge.color}{}

\item{edge.alpha}{}

\item{...}{}
}
\value{
a ggplot2 plot
}
\description{
Twomode graphs
}

