% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_network.R
\name{has.tags}
\alias{has.tags}
\title{Has.tags}
\usage{
has.tags(x, tags, res = "affiliations")
}
\arguments{
\item{x}{an affiliation edge list}

\item{tags}{is a character vector of tags}

\item{res}{if "affiliations" \code{has.tags} returns a character vector with unique affiliation names. If res is "names", a character vector of unique names is returned. If res is "relations", a subset of x is returned.}
}
\description{
Selects all affiliations that have one or several tags
}
\examples{
data(den)
has.tags(den, tags = c("Youth", "Children"))
has.tags(den, tags = c("Youth", "Children"), res = "names")
has.tags(den, tags = c("Youth", "Children"), res = "relations")
}

