% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{hvem}
\alias{hvem}
\title{Hvem er det?}
\usage{
hvem(name, den = "den", only.affiliations = TRUE, ignore.case = TRUE,
  tags = FALSE, ...)
}
\arguments{
\item{name}{the name of the individual}

\item{den}{an affiliation edge list in a similar format to \link{den}, if "den" the den dataset is used.}

\item{only.affiliations}{if TRUE returns the affiliations of the individual}

\item{ignore.case}{if TRUE grep is not case sensitive}

\item{...}{further arguments are passed on to \link{grep}}

\item{if}{TRUE tags are returned}
}
\value{
A matrix with names and affiliation
}
\description{
Returns the affiliation memberships of an individual, or all direct contacts. Names are matched with grep.
}

