% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_network.R
\name{largest.component}
\alias{largest.component}
\title{Largest component}
\usage{
largest.component(graph, cut.off = 1)
}
\arguments{
\item{graph}{a \link{igraph} network object}

\item{cut.off}{is the minimum number of weighted affiliation memberships a component member has to have.}
}
\value{
a graph object
}
\description{
Finds the largest component in a network
}

