% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{line.plot}
\alias{line.plot}
\title{Line plots}
\usage{
line.plot(x, sort = TRUE, var.names = colnames(x), label.x = NULL,
  label.y = NULL)
}
\arguments{
\item{x}{is a vector or a dataframe of vectors}

\item{sort}{if TRUE values are sorted from highest to lowest}

\item{var.names}{a character vector with the variable names}

\item{label.x}{the label for the x axis}

\item{label.y}{the label for the y axis}
}
\value{
a ggplot2 lineplot
}
\description{
Produces a plot with several lines
}
\examples{
data(den)
health.affil  <- has.tags(den, c("Health"))
den.health    <- den[den$AFFILIATION \%in\% health.affil,]
net.health    <- elite.network(den.health)
measures      <- data.frame(degree = degree(net.health, normalized = TRUE), betweenness = betweenness(net.health, normalized = TRUE))
line.plot(measures)
}

