% Generated by roxygen2 (4.0.2): do not edit by hand
\name{report.two.mode}
\alias{report.two.mode}
\title{A report of the first neighborhood of an individiual}
\usage{
report.two.mode(name, n = Inf)
}
\value{
A list of lists - a list of plots, a dataframe with affiliations, a network object
}
\description{
A report of the first neighborhood of an individiual
}

