% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{tag.network}
\alias{tag.network}
\title{Tag network}
\usage{
tag.network(den, plot = TRUE)
}
\arguments{
\item{den}{an affiliation list}

\item{plot}{If TRUE the network is plotted}
}
\value{
a graph object or a plot
}
\description{
Creates and plots a tag network from an affilation list
}

