% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{tile.plot}
\alias{tile.plot}
\title{Tile.plot
Plots a matrix as tile with color according to intensity}
\usage{
tile.plot(adj.mat, max.color.value = 0.8, text.size = 3,
  brewer.set = "Set1")
}
\arguments{
\item{adj.mat}{is the input matrix, with named rows and columns and numerical cells}
}
\value{
a ggplot2 tile plot
}
\description{
Tile.plot
Plots a matrix as tile with color according to intensity
}
\examples{
data(pe13)
mat                  <- data.frame(degree = pe13$Degree...1, reach = pe13$Reach, memberships = pe13$Memberships)
rownames(mat)        <- as.character(pe13$Name)
adj.mat              <- t(as.matrix(mat[1:20,]))
tile.plot(adj.mat)
}

