% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{tile.plot}
\alias{tile.plot}
\title{Tile.plot
Plots a matrix as tile with color according to intensity}
\usage{
tile.plot(adj.mat, max.color.value = median(adj.mat) * 2, text.size = 3)
}
\arguments{
\item{adj.mat}{is the input matrix, with named rows and columns and numerical cells}
}
\value{
a ggplot2 tile plot
}
\description{
Tile.plot
Plots a matrix as tile with color according to intensity
}

