% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{two.mode}
\alias{two.mode}
\title{Create a two-mode network from an affiliation list}
\usage{
two.mode(rel, weighted = TRUE, directed = FALSE, ...)
}
\arguments{
\item{rel}{an affiliation list}

\item{weighted}{If TRUE the graph is weighted}

\item{directed}{If TRUE the graph is directed - does that even make sense?}

\item{...}{further arguments are passed on to \link{graph.incidence}}
}
\description{
Create a two-mode network from an affiliation list
}

