% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vertex.measures}
\alias{vertex.measures}
\title{Vertex descriptives}
\usage{
vertex.measures(net, reach = 2.1)
}
\arguments{
\item{net}{is an \link{igraph} network object}

\item{reach}{is the maximum distance between two individuals for the reach statistic}
}
\value{
a matrix with a lot of descriptives
}
\description{
Descriptive statistics for each vertex
}

